# Use string.Equals over string.Compare in ContentType for improved performance