@lindexi 2024-09-06
Reference: https://github.com/dotnet/wpf/pull/7836

I forgot about the subsequent optimization...